test(test): reorganize integration tests, remove e2e dir
- move global-install from installation/ to infrastructure/ (alongside socket)
- move content-stability and frame-height from e2e/ to ui/content/ and ui/layout/
- delete planviewer e2e test (flaky, covered by focused feature tests)
- fix describe names to follow 'subdirectory feature-name integration' convention
- replace reduce scroll loop with typeKeys in content-stability
- use test.each in frame-height to eliminate redundancy and broken reduce ordering
- replace expect(true).toBe(false) sentinels with expect.unreachable() in global-install
